#200 Ticket with custom fields and attachments
A workaround to allow users to create new tickets that contains both custom fields and attachments. The MultipartSerializer used cannot handle the custom fields and produces output that isn't consistent with what Freshdesk expects to receive. As a workaround, we strip out custom fields from the original payload and serialize ticket with attachments without CF and then manually add them back into the MultiPartFormDataContent we are going to send to Freshdesk